Simple UI of game
button.e - Locates which button the user pressed on the screen and returns the number of that button (buttons are ordered [1  2])
                                                                                                                                                                                                        [3  4]
draw_screen.e and clear_screen.e - both used in creation of the screen should just have to call draw_screen in beginning of code to get the buttons to appear (see UI_test.e)

flash_button - Takes an input of a number of a button (once again 1-4) and then flashes that button on the VGA. (I need to work with the clock on this somehow the flash happens in like a 1/8000 of  sec.

UI_test.e - Tests all these functions and shows how to use them. It draws the screen and then flashes any button that the user presses.